+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [unreleased]
 
+- Centralized output channel.
 - Reload config and restart extension when config changes.
 - Add `initTagsCommand` and `refreshTagsCommand` to `TagsConfig`.
 - Add _Restart Alloglot_ command.